requirements.txt
# Dependências principais
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets e autenticação
gspread>=5.12.0
google-auth>=2.23.0
oauth2client>=4.1.3
gspread-dataframe>=3.3.0

# Utilidades
python-dotenv>=1.0.0
PyYAML>=6.0.0
Pillow>=10.0.0  # Para processamento de imagens

# Cache e desempenho
joblib>=1.3.0  # Para cache em disco
python-multipart>=0.0.6  # Para upload de arquivos

# Acessibilidade
beautifulsoup4>=4.12.0  # Para manipulação de HTML
lxml>=4.9.0  # Parser para HTML/XML

# Testes (opcional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Desenvolvimento
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
